Revert "Remove run_dssat absolute path from sample script"
This reverts commit d2be9d2ee0144fcbf04647467554b8db2ce05810.

This commit was introduced by mistake when preparing a patch for the
Spack installation: so far we are hard-coding run_dssat path in
gym_dssat (both in dssat_pdi.py, run_env.py and run_env_dev.py) as
/opt/dssat_pdi/run_dssat, since that is the DSSAT installation path
when using binary packages. However, the Spack installation is at $HOME,
so a patch is applied during the spack processing to remove that
absolute path (/opt/dssat_pdi) and just leave a PATH relative
'run_dssat'.